+++ sheriff/CHANGELOG.md
@@ -1,5 +1,13 @@
-# Revision history for code-checker
+# Revision history for sheriff
+
+## 0.2.0.1
+* Add derived signature from the arg types for Signature Check rules
+
+## 0.2.0.0
+* Add allowed modules list for function rule
+* Add support for wilcard character "*" in function rules
+* Add support for blocking functions with a particular type signature (Signature only)
+* Add test rules in sheriff rules
 
 ## 0.1.0.0 -- 2024-03-20
-
 * First version. Basic rules based compilation error.